requirements.txt
anthropic==0.84.0
google-genai==1.47.0
greenlet==3.2.5
pytest==8.4.2
numpy==2.0.2
pandas==2.3.3
playwright==1.58.0
pyee==13.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.3
